Address comments - remove utils module
