+++ mlflow/pyfunc/cli.py
@@ -10,8 +10,7 @@ import click
 import pandas
 
 from mlflow.projects import _get_conda_bin_executable, _get_or_create_conda_env
-from mlflow.pyfunc import load_pyfunc, scoring_server
-from mlflow.pyfunc.utils import _load_model_env
+from mlflow.pyfunc import load_pyfunc, scoring_server, _load_model_env
 from mlflow.tracking.utils import _get_model_log_dir
 from mlflow.utils import cli_args
